+++ docs/changes.rst
@@ -15,6 +15,7 @@ v2.3.1 (UNRELEASED)
 - New: ``iamarg``
 - Update: ``lookingforgroup`` after feed change.
 - Update: ``phd`` after feed change.
+- Inactive: ``lunch``
 
 
 v2.3.0 (2014-04-07)